refactor: improve video list and view model
