+++ MindGear_iOS/MindGear_iOS/ViewModels/VideoViewModel.swift
@@ -1,16 +1,24 @@
 import Foundation
 
+/// ViewModel steuert das Laden und Filtern der Videos.
+/// Einfach gehalten, damit der Einstieg in SwiftUI klar bleibt.
+
 @MainActor
 class VideoViewModel: ObservableObject {
     @Published var videos: [Video] = []
     @Published var errorMessage: String? = nil
     @Published var searchText: String = ""
+    @Published var isLoading: Bool = false
 
     // Dynamically load API credentials from Config.plist
     private let apiKey = ConfigManager.apiKey
     private let playlistId = ConfigManager.playlistId
 
+    /// Lädt die Videos der konfigurierten YouTube-Playlist.
+    /// Zeigt einen Ladeindikator und fängt einfache Fehler ab.
     func loadVideos() async {
+        isLoading = true
+        defer { isLoading = false }
         do {
             let items = try await APIService.shared.fetchVideos(from: playlistId, apiKey: apiKey)
             self.videos = items.map { item in
@@ -47,12 +55,16 @@ class VideoViewModel: ObservableObject {
 
     @Published var showFavoritesOnly: Bool = false
 
+    /// Gibt Videos basierend auf Suchbegriff und Favoritenstatus zurück.
     var filteredVideos: [Video] {
+        let trimmed = searchText.trimmingCharacters(in: .whitespacesAndNewlines)
+        let search = trimmed.lowercased()
+
         return videos.filter { video in
-            let matchesSearch = searchText.isEmpty ||
-                video.title.localizedCaseInsensitiveContains(searchText) ||
-                video.description.localizedCaseInsensitiveContains(searchText) ||
-                video.category.localizedCaseInsensitiveContains(searchText)
+            let matchesSearch = search.isEmpty ||
+                video.title.lowercased().contains(search) ||
+                video.description.lowercased().contains(search) ||
+                video.category.lowercased().contains(search)
 
             let matchesFavorites = !showFavoritesOnly || video.isFavorite
